requirements.txt
numpy
mecab-python3
jpype1
nltk
spacy